# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoCango Vehicle Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /favorites:
    get:
      operationId: listFavorites
      tags: [buyer]
      summary: List the authenticated buyer's favorite vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Favorites list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoritesListResponse"
  /favorites/{vehicleId}:
    post:
      operationId: toggleFavorite
      tags: [buyer]
      summary: Toggle a vehicle in/out of favorites
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Toggle result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleFavoriteResponse"

  # ── Conversations ─────────────────────────────────────────────────────────
...
  /freight-conversations:
    get:
      operationId: listFreightConversations
      tags: [freight]
      summary: List freight conversations for the current user (seller or forwarder)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Freight conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreightConversation"

    post:
      operationId: createFreightConversation
      tags: [freight]
      summary: Start a freight conversation between seller and forwarder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFreightConversationInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreightConversation"
  /freight-conversations/{conversationId}:
    get:
      operationId: getFreightConversation
      tags: [freight]
      summary: Get freight conversation detail
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreightConversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /freight-conversations/{conversationId}/messages:
    get:
      operationId: listFreightMessages
      tags: [freight]
      summary: List messages in a freight conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreightMessageListResponse"

    post:
      operationId: sendFreightMessage
      tags: [freight]
      summary: Send a message in a freight conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFreightMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreightMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    VehicleSummary:
      type: object
      required: [id, brandName, modelName, year, fuelType, condition, status, createdAt]
      properties:
        id:
          type: string
        brandName:
          type: string
        modelName:
          type: string
        trim:
          type: ["string", "null"]
        year:
          type: integer
        fuelType:
          type: string
          enum: [petrol, diesel, electric, hybrid, phev, hydrogen, lpg, other]
        condition:
          type: string
          enum: [new, used, certified_used]
        exteriorColor:
          type: ["string", "null"]
        mileageKm:
          type: integer
        displayPriceUsd:
          type: ["number", "null"]
          description: "Final buyer-facing price (FOB + platform commission). Always use this for display; never expose fobPriceUsd to buyers."
        fobPriceUsd:
          type: ["string", "null"]
          description: "Raw seller FOB price. Only included for seller-own and admin views; absent from public buyer responses."
        quantity:
          type: integer
        status:
          type: string
          enum: [draft, pending_review, published, rejected, archived]
        primaryImageUrl:
          type: ["string", "null"]
        viewCount:
          type: integer
        sellerId:
          type: string
        sellerName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    FavoriteItem:
      type: object
      required: [id, vehicleId, createdAt, vehicle]
      properties:
        id:
          type: string
        vehicleId:
          type: string
        createdAt:
          type: string
        vehicle:
          $ref: "#/components/schemas/VehicleSummary"
    FavoritesListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ToggleFavoriteResponse:
      type: object
      required: [isFavorited, vehicleId]
      properties:
        isFavorited:
          type: boolean
        vehicleId:
          type: string

    # ── Chat ─────────────────────────────────────────────────────────────────
...
    FreightConversation:
      type: object
      required: [id, orderId, sellerId, forwarderId, sellerUnread, forwarderUnread, lastMessageAt, createdAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        freightRequestId:
          type: string
          nullable: true
        sellerId:
          type: string
        sellerName:
          type: string
          nullable: true
        forwarderId:
          type: string
        forwarderName:
          type: string
          nullable: true
        sellerUnread:
          type: integer
        forwarderUnread:
          type: integer
        lastMessageAt:
          type: string
          format: date-time
        lastMessagePreview:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    FreightMessage:
      type: object
      required: [id, conversationId, senderId, body, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
          nullable: true
        body:
          type: string
        wasScrubbedAt:
          type: string
          format: date-time
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    FreightMessageListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FreightMessage"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateFreightConversationInput:
      type: object
      required: [orderId, forwarderId]
      properties:
        orderId:
          type: string
        forwarderId:
          type: string
        freightRequestId:
          type: string
          nullable: true
    SendFreightMessageBody:
      type: object
      required: [body]
      properties:
        body:
          type: string

    # ── Admin Schemas ─────────────────────────────────────────────────────────
...